still writing

+++ QClient/sendfile.cpp
@@ -13,9 +13,11 @@ SendFile::~SendFile()
     delete ui;
 }
 
-void SendFile::initSocket(QFile *file, QHostAddress &destination, quint16 destinationPort, QHostAddress &address, quint16 port)
+void SendFile::initSocket(QFile *file,
+                          QHostAddress &destination, quint16 destinationPort,
+                          QHostAddress &address, quint16 port)
 {
-    fileSize = static_cast<quint64>(file->size ());
+    fileSize = file->size ();
     thread = new SendFileThread(this, file, destination, destinationPort, address, port);
     connect (thread, &SendFileThread::updateProgress, this, &SendFile::updateProgress, Qt::QueuedConnection);
     thread->start ();
